refactor(sysadmin): split service.test.ts and remove pure-function class methods
Split the 1,042-line service.test.ts into 4 files matching the new
module boundaries: classifiers.test.ts (8 tests), aggregations.test.ts
(34 tests, 6 describes), pagination.test.ts (8 tests), and a slimmed
service.test.ts (9 orchestration tests for getMonthlyActivity /
getAlerts only). The shared member() factory moved to fixtures.ts;
jest.config.cjs ignores it via the same testPathIgnorePatterns shape
used for helpers.ts.

Test bodies are unchanged. Only the call form switches from
`service.computeXxx(...)` (class method) to `computeXxx(...)` (module
function import). The corresponding 7 wrapper class methods in
SysAdminService and the service.ts re-exports of moved symbols are
removed in the same commit so the test split and class shrink land
atomically.
