+++ jest.config.cjs
@@ -35,6 +35,7 @@ module.exports = {
   testPathIgnorePatterns: [
     "<rootDir>/src/__tests__/helper/*.*",
     "<rootDir>/src/__tests__/.*/helpers\\.ts",
+    "<rootDir>/src/__tests__/.*/fixtures\\.ts",
     "<rootDir>/dist/",
   ],
   moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
